[depixel] revert the previous commit and represent the image size as an integer

+++ src/depixel-cuda/main.cu
@@ -18,7 +18,7 @@ int main(int argc, char** argv) {
   int height = atoi(argv[2]);
   int repeat = atoi(argv[3]);
 
-  size_t size = (size_t)width * height;
+  int size = width * height;
   size_t size_output_bytes = size * sizeof(uint);
   size_t size_image_bytes = size * sizeof(float3);
 
@@ -46,7 +46,7 @@ int main(int argc, char** argv) {
 
   for (int n = 0; n < repeat; n++) {
 
-    for (size_t i = 0; i < size; i++) {
+    for (int i = 0; i < size; i++) {
       h_img[i].x = dis(gen);
       h_img[i].y = dis(gen);
       h_img[i].z = dis(gen);
@@ -68,7 +68,7 @@ int main(int argc, char** argv) {
     cudaMemcpy(h_out, d_out, size_output_bytes, cudaMemcpyDeviceToHost);
 
     float lsum = 0;
-    for (size_t i = 0; i < size; i++)
+    for (int i = 0; i < size; i++)
       lsum += (h_out[i] & 0xff) / 256.f + 
              ((h_out[i] >> 8) & 0xff) / 256.f + 
              ((h_out[i] >> 16) & 0xff) / 256.f +
